+++ jest-puppeteer.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  exitOnPageError: false,
+  launch: {
+    headless: process.env.HEADLESS !== 'false',
+    args: ['--start-maximized'],
+  },
+};