+++ src/models/Contest.ts
@@ -1,4 +1,5 @@
 import { Sendable } from './Sendable';
+import { Config } from '../utils/Config';
 
 export class Contest implements Sendable {
   constructor(public tasks: Sendable[]) {
@@ -6,11 +7,22 @@ export class Contest implements Sendable {
 
   send(): Promise<void> {
     return new Promise(async (resolve) => {
-      for (let i = 0; i < this.tasks.length; i++) {
-        await this.tasks[i].send();
-      }
+      Config.get<boolean>('debugMode').then(async debug => {
+        if (debug) {
+          console.log(JSON.stringify(this.tasks, null, 4));
+        }
 
-      resolve();
+        const oldLog = console.log;
+        console.log = () => {};
+
+        for (let i = 0; i < this.tasks.length; i++) {
+          await this.tasks[i].send();
+        }
+
+        console.log = oldLog;
+
+        resolve();
+      }).catch(console.error);
     });
   }
 }