+++ src/parsers/problem/TimusProblemParser.ts
@@ -29,13 +29,24 @@ export class TimusProblemParser extends Parser {
 
       task.setGroup(group);
 
-      [...elem.querySelectorAll('.sample tbody tr')].slice(1).forEach(tr => {
-        const columns = tr.querySelectorAll('td');
-        const input = columns[0].textContent.trim();
-        const output = columns[1].textContent.trim();
+      if (elem.querySelector('.sample tbody tr > td:nth-child(2)') !== null) {
+        [...elem.querySelectorAll('.sample tbody tr')].slice(1).forEach(tr => {
+          const columns = tr.querySelectorAll('td');
+          const input = columns[0].textContent.trim();
+          const output = columns[1].textContent.trim();
 
-        task.addTest(new Test(input, output));
-      });
+          task.addTest(new Test(input, output));
+        });
+      } else {
+        const blocks = [...elem.querySelectorAll('.sample tbody tr pre')];
+
+        for (let i = 0; i < blocks.length; i += 2) {
+          const input = blocks[i].textContent.trim();
+          const output = blocks[i + 1].textContent.trim();
+
+          task.addTest(new Test(input, output));
+        }
+      }
 
       resolve(task.build());
     });