+++ src/utils/dom.ts
@@ -1,10 +1,14 @@
 // This package doesn't have types
-const snarkdown = require('snarkdown').default;
+const snarkdown = require('snarkdown');
 
 export function htmlToElement(html: string): Element {
   return new DOMParser().parseFromString(html, 'text/html').body;
 }
 
 export function markdownToHtml(markdown: string): string {
+  if (snarkdown.default) {
+    return snarkdown.default(markdown);
+  }
+
   return snarkdown(markdown);
 }